# src/dental_practice_admin/principle_openapi.json
{
  "components": {
    "parameters": {
...
      "practiceId": {
        "description": "",
        "in": "path",
        "name": "practiceId",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "NotFoundError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Response"
      },
      "UnauthorisedError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Response",
        "headers": {
          "WWW-Authenticate": {
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "description": "",
            "type": "string"
          },
          "message": {
            "description": "",
            "type": "string"
          },
          "statusCode": {
            "description": "",
            "type": "integer"
          }
        },
        "required": [
          "error",
          "statusCode"
        ],
        "type": "object"
      },
...
      "Practice": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "oauth2": {
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.principle.dental/oauth/authorize",
            "refreshUrl": "https://api.principle.dental/oauth/refresh",
            "scopes": {
              "accounts:read": "Read account data including invoices / transactions / deposits",
              "accounts:write": "Write account data including invoices / transactions / deposits",
              "appointments:read": "Read appointment data including online booking requests",
              "appointments:write": "Write appointment data including online booking requests",
              "availability:read": "Read availability data including appointment search helpers",
              "emails:read": "Read email data",
              "emails:write": "Write email data",
              "files:read": "Read patient file data",
              "files:write": "Write patient file data",
              "interactions:write": "Write interaction data for patients / appointments",
              "messages:read": "Read message data",
              "messages:write": "Write message data",
              "notifications:write": "Write notification data",
              "patients:read": "Read patient data",
              "patients:write": "Write patient data",
              "practices:read": "Read practice data",
              "tags:read": "Read tag data",
              "tags:write": "Write tag data",
              "treatment-plans:read": "Read treatment plans, steps, and treatments",
              "treatment-plans:write": "Write treatment plans, steps, and treatments"
            },
            "tokenUrl": "https://api.principle.dental/oauth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Principle interface",
    "version": "released"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/practices": {
      "get": {
        "operationId": "listPractices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/Practice"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorisedError"
          }
        },
        "security": [
          {
            "oauth2": [
              "availability:read",
              "appointments:read",
              "appointments:write",
              "practices:read"
            ]
          },
          {
            "apiKeyAuth": []
          }
        ],
        "tags": [
          "Practices"
        ]
      }
    },
    "/v1/practices/{practiceId}": {
      "get": {
        "operationId": "getPractice",
        "parameters": [
          {
            "$ref": "#/components/parameters/practiceId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Practice"
                }
              }
            },
            "description": "Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorisedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        },
        "security": [
          {
            "oauth2": [
              "practices:read"
            ]
          },
          {
            "apiKeyAuth": []
          }
        ],
        "tags": [
          "Practices"
        ]
      }
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    },
    {
      "oauth2": []
    }
  ],
...
}